Speed speed! Max speed. :+1:

+++ src/main/java/onlinemarket/dao/ProvinceDao.java
@@ -2,6 +2,6 @@ package onlinemarket.dao;
 
 import onlinemarket.model.Province;
 
-public interface ProvinceDao extends InterfaceDao<Byte, Province>{
+public interface ProvinceDao extends InterfaceDao<Integer, Province>{
 
 }
