+++ src/main/java/onlinemarket/service/ProvinceService.java
@@ -4,6 +4,6 @@ import onlinemarket.form.filter.FilterForm;
 import onlinemarket.model.Province;
 import onlinemarket.result.ResultObject;
 
-public interface ProvinceService extends InterfaceService<Byte, Province> {
+public interface ProvinceService extends InterfaceService<Integer, Province> {
 	public ResultObject<Province> list(FilterForm filterForm);
 }
